Fixed bug. Update config

+++ common/data.py
@@ -146,7 +146,7 @@ class CoRL2017Dataset(object):
         slidingWindow = self.slidingWindow
         k1 = sequence_len* int( (framePerFile-sequence_len)/slidingWindow + 1 )
         k2 = sequence_len-slidingWindow
-        arr = np.array([ slidingWindow*x + int(x/k1)*k2 for x in arr ])
+        arr = np.array([ slidingWindow*x + int(x/k1)*k2 for x in arr ]).astype(int)
 
         # Expand
         IDs = [ np.array(range(idx,idx+sequence_len)) for idx in arr ]
